% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popcount.R
\name{popcount}
\alias{popcount}
\title{Count populations (# of 1-bits) of integers}
\usage{
popcount(xs)
}
\arguments{
\item{xs}{An integer vector}
}
\value{
The population (# of 1-bits) of each integer
}
\description{
Count populations (# of 1-bits) of integers
}
